+++ src/main/java/fll/scheduler/ScheduleWriter.java
@@ -149,37 +149,37 @@ public final class ScheduleWriter {
     final Element teamNumberHeaderCell = FOPUtils.createTableCell(document, null,
                                                                   TournamentSchedule.TEAM_NUMBER_HEADER);
     headerRow.appendChild(teamNumberHeaderCell);
-    FOPUtils.addBorders(teamNumberHeaderCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                        STANDARD_BORDER_WIDTH);
+    FOPUtils.addBorders(teamNumberHeaderCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                        FOPUtils.STANDARD_BORDER_WIDTH);
 
     final Element teamNameHeaderCell = FOPUtils.createTableCell(document, null, TournamentSchedule.TEAM_NAME_HEADER);
     headerRow.appendChild(teamNameHeaderCell);
-    FOPUtils.addBorders(teamNameHeaderCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                        STANDARD_BORDER_WIDTH);
+    FOPUtils.addBorders(teamNameHeaderCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                        FOPUtils.STANDARD_BORDER_WIDTH);
 
     final Element organizationHeaderCell = FOPUtils.createTableCell(document, null,
                                                                     TournamentSchedule.ORGANIZATION_HEADER);
     headerRow.appendChild(organizationHeaderCell);
-    FOPUtils.addBorders(organizationHeaderCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                        STANDARD_BORDER_WIDTH);
+    FOPUtils.addBorders(organizationHeaderCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                        FOPUtils.STANDARD_BORDER_WIDTH);
 
     final Element judgeGroupHeaderCell = FOPUtils.createTableCell(document, null,
                                                                   TournamentSchedule.JUDGE_GROUP_HEADER);
     headerRow.appendChild(judgeGroupHeaderCell);
-    FOPUtils.addBorders(judgeGroupHeaderCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                        STANDARD_BORDER_WIDTH);
+    FOPUtils.addBorders(judgeGroupHeaderCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                        FOPUtils.STANDARD_BORDER_WIDTH);
 
     final Element awardGroupHeaderCell = FOPUtils.createTableCell(document, null,
                                                                   TournamentSchedule.AWARD_GROUP_HEADER);
     headerRow.appendChild(awardGroupHeaderCell);
-    FOPUtils.addBorders(awardGroupHeaderCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                        STANDARD_BORDER_WIDTH);
+    FOPUtils.addBorders(awardGroupHeaderCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                        FOPUtils.STANDARD_BORDER_WIDTH);
 
     for (final String subjectiveStation : subjectiveStations) {
       final Element subjHeaderCell = FOPUtils.createTableCell(document, null, subjectiveStation);
       headerRow.appendChild(subjHeaderCell);
-      FOPUtils.addBorders(subjHeaderCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(subjHeaderCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
     }
     for (int round = 0; round < schedule.getNumberOfPracticeRounds(); ++round) {
       final Element headerCell1 = FOPUtils.createTableCell(document, null,
@@ -187,31 +187,31 @@ public final class ScheduleWriter {
                                                                          round
                                                                              + 1));
       headerRow.appendChild(headerCell1);
-      FOPUtils.addBorders(headerCell1, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(headerCell1, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       final Element headerCell2 = FOPUtils.createTableCell(document, null,
                                                            String.format(TournamentSchedule.PRACTICE_TABLE_HEADER_FORMAT,
                                                                          round
                                                                              + 1));
       headerRow.appendChild(headerCell2);
-      FOPUtils.addBorders(headerCell2, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(headerCell2, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
     }
     for (int round = 0; round < schedule.getNumberOfRegularMatchPlayRounds(); ++round) {
       final Element headerCell1 = FOPUtils.createTableCell(document, null,
                                                            String.format(TournamentSchedule.PERF_HEADER_FORMAT, round
                                                                + 1));
       headerRow.appendChild(headerCell1);
-      FOPUtils.addBorders(headerCell1, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(headerCell1, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       final Element headerCell2 = FOPUtils.createTableCell(document, null,
                                                            String.format(TournamentSchedule.TABLE_HEADER_FORMAT, round
                                                                + 1));
       headerRow.appendChild(headerCell2);
-      FOPUtils.addBorders(headerCell2, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(headerCell2, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
     }
 
     final Element tableBody = FOPUtils.createXslFoElement(document, FOPUtils.TABLE_BODY_TAG);
@@ -225,28 +225,28 @@ public final class ScheduleWriter {
 
       final Element teamNumberCell = FOPUtils.createTableCell(document, null, String.valueOf(si.getTeamNumber()));
       row.appendChild(teamNumberCell);
-      FOPUtils.addBorders(teamNumberCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(teamNumberCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       final Element teamNameCell = FOPUtils.createTableCell(document, null, si.getTeamName());
       row.appendChild(teamNameCell);
-      FOPUtils.addBorders(teamNameCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(teamNameCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       final Element organizationCell = FOPUtils.createTableCell(document, null, si.getOrganization());
       row.appendChild(organizationCell);
-      FOPUtils.addBorders(organizationCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(organizationCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       final Element judgeGroupCell = FOPUtils.createTableCell(document, null, si.getJudgingGroup());
       row.appendChild(judgeGroupCell);
-      FOPUtils.addBorders(judgeGroupCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(judgeGroupCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       final Element awardGroupCell = FOPUtils.createTableCell(document, null, si.getAwardGroup());
       row.appendChild(awardGroupCell);
-      FOPUtils.addBorders(awardGroupCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(awardGroupCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       for (final String subjectiveStation : subjectiveStations) {
         final SubjectiveTime stime = si.getSubjectiveTimeByName(subjectiveStation);
@@ -256,8 +256,8 @@ public final class ScheduleWriter {
         }
         final Element cell = FOPUtils.createTableCell(document, null, TournamentSchedule.formatTime(stime.getTime()));
         row.appendChild(cell);
-        FOPUtils.addBorders(cell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                            STANDARD_BORDER_WIDTH);
+        FOPUtils.addBorders(cell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                            FOPUtils.STANDARD_BORDER_WIDTH);
       }
 
       si.enumeratePracticePerformances().forEachOrdered(Errors.rethrow().wrap(pair -> {
@@ -265,14 +265,14 @@ public final class ScheduleWriter {
 
         final Element cell1 = FOPUtils.createTableCell(document, null, TournamentSchedule.formatTime(perf.getTime()));
         row.appendChild(cell1);
-        FOPUtils.addBorders(cell1, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                            STANDARD_BORDER_WIDTH);
+        FOPUtils.addBorders(cell1, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                            FOPUtils.STANDARD_BORDER_WIDTH);
 
         final Element cell2 = FOPUtils.createTableCell(document, null,
                                                        String.format("%s %s", perf.getTable(), perf.getSide()));
         row.appendChild(cell2);
-        FOPUtils.addBorders(cell2, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                            STANDARD_BORDER_WIDTH);
+        FOPUtils.addBorders(cell2, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                            FOPUtils.STANDARD_BORDER_WIDTH);
       }));
 
       si.enumerateRegularMatchPlayPerformances().forEachOrdered(Errors.rethrow().wrap(pair -> {
@@ -280,14 +280,14 @@ public final class ScheduleWriter {
 
         final Element cell1 = FOPUtils.createTableCell(document, null, TournamentSchedule.formatTime(perf.getTime()));
         row.appendChild(cell1);
-        FOPUtils.addBorders(cell1, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                            STANDARD_BORDER_WIDTH);
+        FOPUtils.addBorders(cell1, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                            FOPUtils.STANDARD_BORDER_WIDTH);
 
         final Element cell2 = FOPUtils.createTableCell(document, null,
                                                        String.format("%s %s", perf.getTable(), perf.getSide()));
         row.appendChild(cell2);
-        FOPUtils.addBorders(cell2, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                            STANDARD_BORDER_WIDTH);
+        FOPUtils.addBorders(cell2, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                            FOPUtils.STANDARD_BORDER_WIDTH);
       }));
 
     }
@@ -355,16 +355,6 @@ public final class ScheduleWriter {
       throw new FLLInternalException("Error creating the performance schedule PDF", e);
     }
   }
-
-  /**
-   * Standard border for table cells in points.
-   */
-  public static final double STANDARD_BORDER_WIDTH = 0.5;
-
-  /**
-   * Think border for table cells in points.
-   */
-  public static final double THICK_BORDER_WIDTH = 2;
 
   private static Document createPerformanceSchedulePerTable(final TournamentSchedule schedule,
                                                             final boolean displayAwardGroup,
@@ -526,10 +516,10 @@ public final class ScheduleWriter {
     final Element tournamentHeader = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, headerText);
     headerRow1.appendChild(tournamentHeader);
     tournamentHeader.setAttribute("number-columns-spanned", String.valueOf(headerNames.size()));
-    FOPUtils.addTopBorder(tournamentHeader, STANDARD_BORDER_WIDTH);
-    FOPUtils.addBottomBorder(tournamentHeader, STANDARD_BORDER_WIDTH);
-    FOPUtils.addLeftBorder(tournamentHeader, STANDARD_BORDER_WIDTH);
-    FOPUtils.addRightBorder(tournamentHeader, STANDARD_BORDER_WIDTH);
+    FOPUtils.addTopBorder(tournamentHeader, FOPUtils.STANDARD_BORDER_WIDTH);
+    FOPUtils.addBottomBorder(tournamentHeader, FOPUtils.STANDARD_BORDER_WIDTH);
+    FOPUtils.addLeftBorder(tournamentHeader, FOPUtils.STANDARD_BORDER_WIDTH);
+    FOPUtils.addRightBorder(tournamentHeader, FOPUtils.STANDARD_BORDER_WIDTH);
 
     final Element headerRow2 = FOPUtils.createTableRow(document);
     header.appendChild(headerRow2);
@@ -537,8 +527,8 @@ public final class ScheduleWriter {
     for (final String headerName : headerNames) {
       final Element headerCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, headerName);
       headerRow2.appendChild(headerCell);
-      FOPUtils.addBorders(headerCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(headerCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
     }
 
     final Element tableBody = FOPUtils.createXslFoElement(document, FOPUtils.TABLE_BODY_TAG);
@@ -567,66 +557,66 @@ public final class ScheduleWriter {
       final LocalTime performanceTime = performance.getTime();
       final double topBorderWidth;
       if (Objects.equals(performanceTime, prevTime)) {
-        topBorderWidth = STANDARD_BORDER_WIDTH;
+        topBorderWidth = FOPUtils.STANDARD_BORDER_WIDTH;
 
         // keep the rows with the same times together
         FOPUtils.keepWithPrevious(row);
       } else {
-        topBorderWidth = THICK_BORDER_WIDTH;
+        topBorderWidth = FOPUtils.THICK_BORDER_WIDTH;
       }
 
       final Element teamNumberCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
                                                               String.valueOf(si.getTeamNumber()));
       row.appendChild(teamNumberCell);
-      FOPUtils.addBorders(teamNumberCell, topBorderWidth, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(teamNumberCell, topBorderWidth, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       if (displayAwardGroup) {
         final Element awardGroupCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
                                                                 String.valueOf(si.getAwardGroup()));
         row.appendChild(awardGroupCell);
-        FOPUtils.addBorders(awardGroupCell, topBorderWidth, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                            STANDARD_BORDER_WIDTH);
+        FOPUtils.addBorders(awardGroupCell, topBorderWidth, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                            FOPUtils.STANDARD_BORDER_WIDTH);
       }
 
       final Element organizationCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
                                                                 String.valueOf(si.getOrganization()));
       row.appendChild(organizationCell);
-      FOPUtils.addBorders(organizationCell, topBorderWidth, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(organizationCell, topBorderWidth, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       final Element teamNameCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
                                                             String.valueOf(si.getTeamName()));
       row.appendChild(teamNameCell);
-      FOPUtils.addBorders(teamNameCell, topBorderWidth, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(teamNameCell, topBorderWidth, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       final Element timeCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
                                                         TournamentSchedule.formatTime(performanceTime));
       row.appendChild(timeCell);
-      FOPUtils.addBorders(timeCell, topBorderWidth, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(timeCell, topBorderWidth, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       final Element perfTableCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
                                                              performance.getTable()
                                                                  + " "
                                                                  + performance.getSide());
       row.appendChild(perfTableCell);
-      FOPUtils.addBorders(perfTableCell, topBorderWidth, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(perfTableCell, topBorderWidth, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       final Element perfRoundCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
                                                              si.getRoundName(performance));
       row.appendChild(perfRoundCell);
-      FOPUtils.addBorders(perfRoundCell, topBorderWidth, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(perfRoundCell, topBorderWidth, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
 
       if (displayNotes) {
         final Element notesCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
                                                            String.valueOf(Utilities.NON_BREAKING_SPACE));
         row.appendChild(notesCell);
-        FOPUtils.addBorders(notesCell, topBorderWidth, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                            STANDARD_BORDER_WIDTH);
+        FOPUtils.addBorders(notesCell, topBorderWidth, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                            FOPUtils.STANDARD_BORDER_WIDTH);
       }
 
       prevTime = performanceTime;
@@ -920,8 +910,8 @@ public final class ScheduleWriter {
                                                                             subjectiveStation));
     headerRow1.appendChild(tournamentHeader);
     tournamentHeader.setAttribute("number-columns-spanned", String.valueOf(headerNames.length));
-    FOPUtils.addBorders(tournamentHeader, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                        STANDARD_BORDER_WIDTH);
+    FOPUtils.addBorders(tournamentHeader, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                        FOPUtils.STANDARD_BORDER_WIDTH);
 
     final Element headerRow2 = FOPUtils.createTableRow(document);
     header.appendChild(headerRow2);
@@ -929,8 +919,8 @@ public final class ScheduleWriter {
     for (final String headerName : headerNames) {
       final Element headerCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, headerName);
       headerRow2.appendChild(headerCell);
-      FOPUtils.addBorders(headerCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(headerCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
     }
 
     final Element tableBody = FOPUtils.createXslFoElement(document, FOPUtils.TABLE_BODY_TAG);
@@ -953,7 +943,7 @@ public final class ScheduleWriter {
 
       final double topBorderWidth;
       if (Objects.equals(time, prevTime)) {
-        topBorderWidth = STANDARD_BORDER_WIDTH;
+        topBorderWidth = FOPUtils.STANDARD_BORDER_WIDTH;
 
         // keep the rows with the same times together
         FOPUtils.keepWithPrevious(row);
@@ -963,44 +953,44 @@ public final class ScheduleWriter {
 
       Element cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, String.valueOf(si.getTeamNumber()));
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, si.getAwardGroup());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, si.getOrganization());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, si.getTeamName());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, TournamentSchedule.formatTime(time));
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, si.getJudgingGroup());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       prevTime = time;
@@ -1087,8 +1077,8 @@ public final class ScheduleWriter {
     final Element tournamentHeader = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, headerFooterText);
     headerRow1.appendChild(tournamentHeader);
     tournamentHeader.setAttribute("number-columns-spanned", String.valueOf(headerNames.length));
-    FOPUtils.addBorders(tournamentHeader, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                        STANDARD_BORDER_WIDTH);
+    FOPUtils.addBorders(tournamentHeader, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                        FOPUtils.STANDARD_BORDER_WIDTH);
 
     final Element headerRow2 = FOPUtils.createTableRow(document);
     header.appendChild(headerRow2);
@@ -1096,8 +1086,8 @@ public final class ScheduleWriter {
     for (final String headerName : headerNames) {
       final Element headerCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, headerName);
       headerRow2.appendChild(headerCell);
-      FOPUtils.addBorders(headerCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(headerCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
     }
 
     final Element tableBody = FOPUtils.createXslFoElement(document, FOPUtils.TABLE_BODY_TAG);
@@ -1121,62 +1111,62 @@ public final class ScheduleWriter {
 
       final double topBorderWidth;
       if (Objects.equals(time, prevTime)) {
-        topBorderWidth = STANDARD_BORDER_WIDTH;
+        topBorderWidth = FOPUtils.STANDARD_BORDER_WIDTH;
 
         // keep the rows with the same times together
         FOPUtils.keepWithPrevious(row);
       } else {
-        topBorderWidth = THICK_BORDER_WIDTH;
+        topBorderWidth = FOPUtils.THICK_BORDER_WIDTH;
       }
 
       Element cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
                                               String.valueOf(teamAtTime.getTeamInfo().getTeamNumber()));
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, teamAtTime.getTeamInfo().getAwardGroup());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, teamAtTime.getTeamInfo().getOrganization());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, teamAtTime.getTeamInfo().getTeamName());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, teamAtTime.getSubjTime().getName());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, teamAtTime.getTeamInfo().getJudgingGroup());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, TournamentSchedule.formatTime(time));
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       prevTime = time;
@@ -1274,8 +1264,8 @@ public final class ScheduleWriter {
                                                                             subjectiveStation));
     headerRow1.appendChild(tournamentHeader);
     tournamentHeader.setAttribute("number-columns-spanned", String.valueOf(headerNames.length));
-    FOPUtils.addBorders(tournamentHeader, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                        STANDARD_BORDER_WIDTH);
+    FOPUtils.addBorders(tournamentHeader, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                        FOPUtils.STANDARD_BORDER_WIDTH);
 
     final Element headerRow2 = FOPUtils.createTableRow(document);
     header.appendChild(headerRow2);
@@ -1283,8 +1273,8 @@ public final class ScheduleWriter {
     for (final String headerName : headerNames) {
       final Element headerCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, headerName);
       headerRow2.appendChild(headerCell);
-      FOPUtils.addBorders(headerCell, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH, STANDARD_BORDER_WIDTH,
-                          STANDARD_BORDER_WIDTH);
+      FOPUtils.addBorders(headerCell, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH, FOPUtils.STANDARD_BORDER_WIDTH,
+                          FOPUtils.STANDARD_BORDER_WIDTH);
     }
 
     final Element tableBody = FOPUtils.createXslFoElement(document, FOPUtils.TABLE_BODY_TAG);
@@ -1302,54 +1292,54 @@ public final class ScheduleWriter {
       final String awardGroup = si.getAwardGroup();
       final double topBorderWidth;
       if (Objects.equals(awardGroup, prevAwardGroup)) {
-        topBorderWidth = STANDARD_BORDER_WIDTH;
+        topBorderWidth = FOPUtils.STANDARD_BORDER_WIDTH;
         // keep the rows with the same award group together if possible
         FOPUtils.keepWithPrevious(row);
       } else {
-        topBorderWidth = THICK_BORDER_WIDTH;
+        topBorderWidth = FOPUtils.THICK_BORDER_WIDTH;
       }
 
       Element cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, String.valueOf(si.getTeamNumber()));
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, awardGroup);
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, si.getOrganization());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, si.getTeamName());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
                                       null == stime ? "" : TournamentSchedule.formatTime(stime.getTime()));
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       cell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER, si.getJudgingGroup());
       row.appendChild(cell);
-      FOPUtils.addBottomBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addLeftBorder(cell, STANDARD_BORDER_WIDTH);
-      FOPUtils.addRightBorder(cell, STANDARD_BORDER_WIDTH);
+      FOPUtils.addBottomBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addLeftBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
+      FOPUtils.addRightBorder(cell, FOPUtils.STANDARD_BORDER_WIDTH);
       FOPUtils.addTopBorder(cell, topBorderWidth);
 
       prevAwardGroup = awardGroup;